third_party/blink/web_tests/svg/filters/feBlend-invalid-mode-expected.txt

CONSOLE ERROR: Uncaught TypeError: Failed to set the 'baseVal' property on 'SVGAnimatedEnumeration': The enumeration value provided (65535) is larger than the largest allowed value (16).
PASS: Invalid blend modes do not trigger a crash. 
